added transformString() function to encapsulate multi-byte encoding transformation.
The function is for multibyte users not to be confused on how to prepare DOMDocument
in their environment.

For the case of mbstring is missing, we still keep processing by using meta tags
as a fallback, which may or may not work. The fallback can be found by setting the
log level as debug.

By using HTML entities, a geotag script source may contain HTML entities so it
should be imported as a CDATA node instead of a Text node.