convert &amp;{entitityname}; in output to &{entityname};
All classes extending Element make use of createTextNode which converts symbols
to their respective HTML entities e.g. & -> &amp;

The inputs are not sanitized and therefore if they contain entities the & part
will be converted to &amp; e.g. &quot; -> &amp;quot;

This commit adds a test to prove this and fixes the problem by converting the
render output back.